+++ CHANGELOG.rst
@@ -14,8 +14,9 @@ Version 0.15
 
 - Deprecated syntax for configuration values is now completely rejected. All
   values now have to be valid JSON.
-- The default of ``--list`` for ``vdirsyncer discover`` has changed to
-  ``false``. See :ghpr:`543`.
+- A few UX improvements for Google storages, see :gh:`549` and :gh:`552`.
+- Fix collection discovery for :storage:`google_contacts`, see :gh:`564`.
+- iCloud is now tested on Travis, see :gh:`567`.
 
 Version 0.14.1
 ==============